renamed module app from parametrization to snapshots
