+++ services/web/server/src/simcore_service_webserver/snapshots_settings.py
