+++ services/web/server/tests/unit/isolated/test_snapshots_models.py
@@ -1,5 +1,5 @@
 from simcore_postgres_database.models.snapshots import snapshots
-from simcore_service_webserver.parametrization_models import (
+from simcore_service_webserver.snapshots_models import (
     Parameter,
     ParameterApiModel,
     Snapshot,
