+++ services/web/server/tests/unit/with_dbs/11/test_snapshots_core.py
@@ -7,14 +7,14 @@ from uuid import UUID, uuid4
 
 import pytest
 from aiohttp import web
-from models_library.projects import Project, ProjectAtDB
+from models_library.projects import Project  # , ProjectAtDB
 from models_library.projects_nodes import Node
 from models_library.projects_nodes_io import NodeID
 from simcore_service_webserver.constants import APP_PROJECT_DBAPI
-from simcore_service_webserver.parametrization_core import snapshot_project
 from simcore_service_webserver.projects.projects_api import get_project_for_user
 from simcore_service_webserver.projects.projects_db import APP_PROJECT_DBAPI
 from simcore_service_webserver.projects.projects_utils import clone_project_document
+from simcore_service_webserver.snapshots_core import snapshot_project
 
 
 # is parametrized project?
